Fix hanging HTTP requests
Asynchronous HTTP requests were hanging after completing successfully. This
appears to be caused by a deadlock that occurs when waiting on threads under
certain conditions. The issue is described here: https://stackoverflow.com/a/10369275
and at links from that answer.

This change adds the .ConfigureAwait(false) to all uses of await in the library.